Assignment3/requirements.txt
torch
diffusers
transformers
opencv-python
numpy
Pillow
matplotlib
controlnet_aux